+++ apps/riak_cs/src/riak_cs_list_objects_utils.erl
@@ -46,7 +46,7 @@
 %%%===================================================================
 
 %% API
--export([start_link/5,
+-export([start_link/2,
          get_object_list/1,
          get_internal_state/1]).
 
@@ -58,33 +58,19 @@
          filter_prefix_keys/2,
          extract_group/2]).
 
-%% Observability / Configuration
--export([get_key_list_multiplier/0,
-         set_key_list_multiplier/1,
-         fold_objects_for_list_keys/0,
-         fold_objects_timeout/0]).
-
 %%%===================================================================
 %%% API
 %%%===================================================================
 
 
--spec start_link(riak_client(), pid(), list_object_request(), term(),
-                 UseCache :: boolean()) ->
+-spec start_link(riak_client(), list_object_request()) ->
     {ok, pid()} | {error, term()}.
 %% @doc An abstraction between the old and new list-keys mechanism. Uses the
 %% old mechanism if `fold_objects_for_list_keys' is false, otherwise uses
 %% the new one. After getting a pid back, the API is the same, so users don't
 %% need to differentiate.
-start_link(RcPid, CallerPid, ListKeysRequest, CacheKey, UseCache) ->
-    case fold_objects_for_list_keys() of
-        true ->
-            riak_cs_list_objects_fsm_v2:start_link(RcPid, ListKeysRequest);
-        false ->
-            riak_cs_list_objects_fsm:start_link(RcPid, CallerPid,
-                                                ListKeysRequest, CacheKey,
-                                                UseCache)
-    end.
+start_link(RcPid, ListKeysRequest) ->
+    riak_cs_list_objects_fsm_v2:start_link(RcPid, ListKeysRequest).
 
 -spec get_object_list(pid()) ->
     {ok, list_objects_response() | list_object_versions_response()} |
@@ -214,22 +200,6 @@ update_keys_and_prefixes({ManifestList, Prefixes},
 %%%===================================================================
 %%% Observability / Configuration
 %%%===================================================================
-
--spec get_key_list_multiplier() -> float().
-get_key_list_multiplier() ->
-    riak_cs_config:get_env(riak_cs, key_list_multiplier,
-                          ?KEY_LIST_MULTIPLIER).
-
--spec set_key_list_multiplier(float()) -> 'ok'.
-set_key_list_multiplier(Multiplier) ->
-    application:set_env(riak_cs, key_list_multiplier,
-                        Multiplier).
-
-
--spec fold_objects_for_list_keys() -> boolean().
-fold_objects_for_list_keys() ->
-    riak_cs_config:get_env(riak_cs, fold_objects_for_list_keys,
-                          ?FOLD_OBJECTS_FOR_LIST_KEYS).
 
 -spec fold_objects_timeout() -> non_neg_integer().
 fold_objects_timeout() ->